video upload and conversion real time progress bar using socket

+++ movie-streaming-backend/app.py
@@ -1,7 +1,7 @@
 from flask import Flask, request, jsonify, session
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_socketio import SocketIO, emit
-
+from ffmpeg_progress_yield import FfmpegProgress
 from flask_pymongo import PyMongo
 from flask_cors import CORS
 from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
@@ -13,29 +13,35 @@ from dotenv import load_dotenv
 import mimetypes
 import secrets
 import hashlib
-import subprocess
 from werkzeug.utils import secure_filename
-import os
 import mimetypes
-import re
+
+
 
 
 
 load_dotenv()
 
 app = Flask(__name__)
+
+CORS(app, supports_credentials=True, origins=[
+    "http://localhost:5175",
+    "http://127.0.0.1:5173",
+    "http://172.16.3.135:5173"
+])
+
 socketio = SocketIO(
     app,
     async_mode='eventlet',
     cors_allowed_origins=[
+        "http://localhost:5173",
         "http://127.0.0.1:5173",
         "http://172.16.3.135:5173"
     ]
 )
 
 
-# CORS Configuration - Supporting credentials and allowing specific origins
-CORS(app, supports_credentials=True, origins=["http://127.0.0.1:5173", "http://172.16.3.135:5173"])
+
 
 # Flask Configurations
 app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
@@ -49,7 +55,9 @@ app.config['JWT_TOKEN_LOCATION'] = ['headers']
 app.config["MONGO_URI"] = os.getenv('MONGO_URI')
 
 # Initializing Database and JWT Manager
-db = PyMongo(app).db
+
+mongo = PyMongo(app)
+db = mongo.db
 jwt = JWTManager(app)
 
 MINIO_URL = os.getenv("MINIO_URL")
@@ -69,7 +77,7 @@ if not minio_client.bucket_exists(MINIO_BUCKET_NAME):
 IGNORE_FILES = ['.DS_Store',] 
 
 
-""" def upload_folder_to_minio(localfolder, minioprefix):
+def upload_folder_to_minio(localfolder, minioprefix):
     for foldername, subfolders, filenames in os.walk(localfolder):
         for file in filenames:
             localpath = os.path.join(foldername, file)
@@ -107,71 +115,29 @@ def upload_video_folder():
     minio_path_prefix = f"{title}/"
     upload_folder_to_minio(folder, minio_path_prefix )
 
-    videoUrl = f"http://192.168.101.3:64176/{MINIO_BUCKET_NAME}/{title}/master.m3u8"
+    videoUrl = f"http://127.0.0.1:9000/{MINIO_BUCKET_NAME}/{title}/master.m3u8"
    
     
 
     db.movies.insert_one({
         "title": title,
         "hls_url": videoUrl,
-        "uploaded_at": datetime.datetime.utcnow()
+        "uploaded_at": datetime.now()
     })
 
-    return jsonify({"message": "folder uploaded", "url": videoUrl})   """
- 
+    return jsonify({"message": "folder uploaded", "url": videoUrl})   
 
-#convert into hls 
-@app.route("/admin/convertvideo", methods=["POST"])
-def convert_video():
-    file = request.files['file']
-    title = request.form.get('title', 'untitled')
 
-    filename  = secure_filename(file.filename)
-    upload_path = f"./uploads/{filename}"
-    hls_output_dir = f"./hls_output"
-    minioprefix = f"{title}/"
 
-    os.makedirs(hls_output_dir, exist_ok=True)
-    file.save(upload_path)
-
-   
-  
-
-    command = [
-        "ffmpeg", "-i", upload_path,
-        "-filter_complex", "[0:v]split=3[v1][v2][v3];" +
-                           "[v1]scale=w=1920:h=1080[v1out];" +
-                           "[v2]scale=w=1280:h=720[v2out];" +
-                           "[v3]scale=w=640:h=360[v3out]",
-        "-map", "[v1out]", "-c:v:0", "libx264", "-b:v:0", "5000k",
-        "-map", "[v2out]", "-c:v:1", "libx264", "-b:v:1", "3000k",
-        "-map", "[v3out]", "-c:v:2", "libx264", "-b:v:2", "1000k",
-        "-f", "hls",
-        "-var_stream_map", "v:0,name:1080p v:1,name:720p v:2,name:360p",
-        "-master_pl_name", "master.m3u8",
-        "-hls_time", "4", "-hls_list_size", "0", "-hls_segment_filename",
-        f"{hls_output_dir}/%v/seg_%03d.ts", f"{hls_output_dir}/%v/prog.m3u8"
-    ]
+def ffmpeg_task(command, title, upload_path, hls_output_dir):
+    ff = FfmpegProgress(command)
+    for progress in ff.run_command_with_progress():
+        socketio.emit('progress_update', {'progress': progress})
+        print(progress)
+        socketio.sleep(0)
     
-    process = subprocess.Popen(command, stderr=subprocess.PIPE, universal_newlines=True)
-    duration = None
-
-    for line in process.stderr:
-        if duration is None:
-            match = re.search(r'Duration: (\d+):(\d+):(\d+\.\d+)', line)
-            if match:
-                h, m, s = map(float, match.groups())
-                duration = h * 3600 + m * 60 + s
-
-        match = re.search(r'time=(\d+):(\d+):(\d+\.\d+)', line)
-        if match and duration:
-            h, m, s = map(float, match.groups())
-            current_time = h * 3600 + m * 60 + s
-            percent = int((current_time / duration) * 100)
-            socketio.emit('convert-progress', {'progress': percent})
-
-   
-
+    # Upload files to MinIO after conversion finishes
+    minioprefix = f"{title}/"
     for foldername, subfolders, filenames in os.walk(hls_output_dir):
         for file in filenames:
             localpath = os.path.join(foldername, file)
@@ -179,7 +145,6 @@ def convert_video():
             miniopath = os.path.join(minioprefix, relativepath).replace("\\", "/")
 
             if file in IGNORE_FILES:
-                print(f"Skipping file: {file}")
                 continue
 
             content_type, _ = mimetypes.guess_type(localpath)
@@ -188,24 +153,89 @@ def convert_video():
 
             try:
                 minio_client.fput_object(MINIO_BUCKET_NAME, miniopath, localpath, content_type=content_type)
-                
-                # Permanently delete the file after successful upload``
                 os.remove(localpath)
-                print(f"Deleted local file: {localpath}")
-
             except Exception as e:
                 print(f"Error uploading {file}: {e}")
 
-    os.remove(upload_path)
-    videoUrl = f"http://192.168.101.3:9000/{MINIO_BUCKET_NAME}/{title}/master.m3u8"
+    # Delete the uploaded source file after all done
+    if os.path.exists(upload_path):
+        os.remove(upload_path)
+    
+    try:
+        videoUrl = f"http://192.168.101.2:9000/{MINIO_BUCKET_NAME}/{title}/master.m3u8"
    
-    db.movies.insert_one({
+        db.movies.insert_one({
         "title": title,
         "hls_url": videoUrl,
         "uploaded_at": datetime.now()
-    })
+        })
+    except Exception as e:
+        print(f"failed to save to db {videoUrl}")
 
-    return jsonify({"message": "folder uploaded", "url": videoUrl}) 
+
+
+    socketio.emit('conversion_completed', {'message': f'Conversion for "{title}" completed.', 'url': videoUrl})
+
+        
+
+
+
+#convert into hls 
+@app.route("/admin/convertvideo", methods=["POST"])
+def convert_video():
+    file = request.files['file']
+    title = request.form.get('title', 'untitled')
+   
+
+    filename  = secure_filename(file.filename)
+    BASE_DIR = os.path.abspath(os.path.dirname(__file__))
+    upload_dir = os.path.join(BASE_DIR, 'uploads')
+    upload_path = os.path.join(upload_dir, filename)
+    hls_output_dir = os.path.join(BASE_DIR, 'hls_output')
+    
+    
+    file.save(upload_path)
+    for quality in ['1080p', '720p', '360p']:
+        os.makedirs(os.path.join(hls_output_dir, quality), exist_ok=True)
+
+   
+   
+    socketio.emit('upload_complete', {'message': 'File upload complete', 'filename': filename})
+   
+    
+   
+    command = [
+    "ffmpeg", "-i", upload_path,
+    "-filter_complex", (
+        "[0:v]split=3[v1][v2][v3]; "
+        "[v1]scale=w=1920:h=1080[v1out]; "
+        "[v2]scale=w=854:h=480[v2out]; "
+        "[v3]scale=w=640:h=360[v3out]"
+    ),
+    "-map", "[v1out]", "-c:v:0", "libx264", "-b:v:0", "5000k", "-maxrate:v:0", "5350k", "-bufsize:v:0", "7500k",
+    "-map", "[v2out]", "-c:v:1", "libx264", "-b:v:1", "1400k", "-maxrate:v:1", "1498k", "-bufsize:v:1", "2100k",
+    "-map", "[v3out]", "-c:v:2", "libx264", "-b:v:2", "800k", "-maxrate:v:2", "856k", "-bufsize:v:2", "1200k",
+    "-map", "a:0", "-c:a", "aac", "-b:a:0", "192k", "-ac", "2",
+    "-map", "a:0", "-c:a", "aac", "-b:a:1", "128k", "-ac", "2",
+    "-map", "a:0", "-c:a", "aac", "-b:a:2", "96k", "-ac", "2",
+    "-f", "hls",
+    "-var_stream_map", "v:0,a:0,name:1080p v:1,a:1,name:480p v:2,a:2,name:360p",
+    "-hls_time", "10",
+    "-hls_playlist_type", "vod",
+    "-hls_flags", "independent_segments",
+    "-hls_segment_type", "mpegts",
+    "-hls_segment_filename", f"{hls_output_dir}/stream_%v/data%03d.ts",
+    "-master_pl_name", "master.m3u8",
+   f"{hls_output_dir}/stream_%v/playlist.m3u8"
+]
+
+
+
+    
+
+    socketio.start_background_task(ffmpeg_task, command, title, upload_path, hls_output_dir)
+
+    return jsonify({"message": "Conversion started", "url": None})
 
    
       
@@ -236,6 +266,7 @@ def login():
     data = request.get_json()
     email = data['email']
     password = data['password']
+    print(data)
 
     user = db.users.find_one({"email": email})
     if user and check_password_hash(user['password'], password):
@@ -363,34 +394,13 @@ def protected():
 
 
 
-@socketio.on('connect')
-def handle_connect():
-    print('Client connected')
-   
-@socketio.on('message')
-def handle_my_custom_event(data):
-    print(data)
-    emit('foo_response', {'message': 'got it'})
 
-@socketio.on('create-something')
-def handle_create_something(data):
-    print('🛠️ Received create-something with:', data)
-
-    # Example: process the data (e.g., store in DB, etc.)
-    # ...
-
-    # Acknowledge back to the client
-    return {'status': 'ok', 'message': 'Created successfully'}
-
-@socketio.on('disconnect')
-def handle_disconnect():
-    print('Client disconnected')
 
 
 # Run the Flask Application
 if __name__ == "__main__":
     socketio.run(app, port=5000, debug=True)
-    print(socketio.async_mode)
+
